Suite et fin: MoveTo/CopyTo trait, new module ast::alloc
